requirements.txt
absl-py==2.1.0
aider-chat==0.46.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.1.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.0.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email-validator==1.2.1
executing==2.0.1
Faker==14.2.0
fastjsonschema==2.19.1
feedparser==6.0.10
filelock==3.15.4
flake8==7.1.0
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-sock==0.5.2
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.5.4
gevent==23.9.1
geventhttpclient==2.0.11
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grep-ast==0.3.2
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
ipykernel==6.28.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jproperties==2.1.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.2.0
kiwisolver==1.4.5
langchain==0.0.312
langchain-core==0.2.9
langchain-openai==0.1.9
langcodes==3.3.0
langsmith==0.1.81
libclang==18.1.1
litellm==1.41.27
locust==2.19.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
Movie-Recommendation==0.0.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook_shim==0.2.3
numpy==1.26.4
openai==1.37.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
overrides==7.4.0
packaging==24.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.1.0
playwright==1.45.0
plotly==5.18.0
preshed==3.0.9
prometheus-client==0.19.0
prompt_toolkit==3.0.47
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.1b0
pyee==11.1.0
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypandoc==1.13
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.4
redistimeseries==1.4.5
redisvl==0.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roundrobin==0.0.4
rpds-py==0.19.0
safetensors==0.4.0
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.13.1
Send2Trash==1.8.2
sentence-transformers==3.0.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
thinc==8.2.3
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.1
transformers==4.35.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
validators==0.22.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.1
